Refactor code, introduce NotarytoolNotarizer.
