+++ src/main/java/org/eclipse/cbi/ws/macos/notarization/ApplicationLifecycle.java
@@ -29,29 +29,28 @@ import io.quarkus.runtime.StartupEvent;
 @ApplicationScoped
 public class ApplicationLifecycle {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(ApplicationLifecycle.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(ApplicationLifecycle.class);
 
-  @Inject
-  @ConfigProperty(name = "notarization.cache.uploadedFiles", defaultValue = "/tmp/macos-notarization-service/pending-files")
-  String pendingFiles;
+    @Inject
+    @ConfigProperty(name = "notarization.cache.uploadedFiles", defaultValue = "/tmp/macos-notarization-service/pending-files")
+    String pendingFiles;
 
-  void onStart(@Observes StartupEvent ev) throws IOException {
-    Path pendingFilesPath = Paths.get(pendingFiles);
+    void onStart(@Observes StartupEvent ev) throws IOException {
+        Path pendingFilesPath = Paths.get(pendingFiles);
 
-    if (!Files.isDirectory(pendingFilesPath)) {
-      LOGGER.info("Creating folder '{}'", pendingFilesPath);
-      Files.createDirectories(pendingFilesPath);
-    } else {
-      LOGGER.info("Cleaning up folder '{}'", pendingFilesPath);
-      try (Stream<File> filesToDelete = Files.walk(pendingFilesPath).sorted(Comparator.reverseOrder()).map(Path::toFile)) {
-				filesToDelete.forEach(File::delete);
-			}
-      Files.createDirectories(pendingFilesPath);
+        if (!Files.isDirectory(pendingFilesPath)) {
+            LOGGER.info("Creating folder '{}'", pendingFilesPath);
+            Files.createDirectories(pendingFilesPath);
+        } else {
+            LOGGER.info("Cleaning up folder '{}'", pendingFilesPath);
+            try (Stream<File> filesToDelete = Files.walk(pendingFilesPath).sorted(Comparator.reverseOrder()).map(Path::toFile)) {
+                filesToDelete.forEach(File::delete);
+            }
+            Files.createDirectories(pendingFilesPath);
+        }
     }
-  }
 
-  void onStop(@Observes ShutdownEvent ev) {
-    LOGGER.info("The application is stopping...");
-  }
-
+    void onStop(@Observes ShutdownEvent ev) {
+        LOGGER.info("The application is stopping...");
+    }
 }
