+++ src/main/java/org/eclipse/cbi/ws/macos/notarization/xcrun/common/ComposedStaplerResult.java
@@ -5,7 +5,7 @@
  * which is available at http://www.eclipse.org/legal/epl-v20.html
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.cbi.ws.macos.notarization.xcrun.altool;
+package org.eclipse.cbi.ws.macos.notarization.xcrun.common;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -13,24 +13,24 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 class ComposedStaplerResult implements StaplerResult {
-  private List<StaplerResult> results;
+    private final List<StaplerResult> results;
 
-  ComposedStaplerResult(Collection<? extends StaplerResult> results) {
-    this.results = new ArrayList<>(results);
-  }
-
-  @Override
-  public Status status() {
-    if (results.stream().anyMatch(r -> r.status() == Status.ERROR)) {
-      return Status.ERROR;
+    ComposedStaplerResult(Collection<? extends StaplerResult> results) {
+        this.results = new ArrayList<>(results);
     }
-    return Status.SUCCESS;
-  }
 
-  @Override
-  public String message() {
-    return results.stream()
-      .map(StaplerResult::message)
-      .collect(Collectors.joining("\n"));
-  }
+    @Override
+    public Status status() {
+        if (results.stream().anyMatch(r -> r.status() == Status.ERROR)) {
+            return Status.ERROR;
+        }
+        return Status.SUCCESS;
+    }
+
+    @Override
+    public String message() {
+        return results.stream()
+                .map(StaplerResult::message)
+                .collect(Collectors.joining("\n"));
+    }
 }
