+++ src/main/java/org/eclipse/cbi/ws/macos/notarization/xcrun/common/NotarizationInfoResult.java
@@ -5,7 +5,7 @@
  * which is available at http://www.eclipse.org/legal/epl-v20.html
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.cbi.ws.macos.notarization.xcrun.altool;
+package org.eclipse.cbi.ws.macos.notarization.xcrun.common;
 
 import javax.annotation.Nullable;
 
@@ -18,14 +18,16 @@ public abstract class NotarizationInfoResult {
 	public abstract String message();
 	@Nullable
 	public abstract String notarizationLog();
-	static NotarizationInfoResult.Builder builder() {
+
+	public static NotarizationInfoResult.Builder builder() {
 		return new AutoValue_NotarizationInfoResult.Builder();
 	}
+
 	@AutoValue.Builder
-	static abstract class Builder {
-		abstract NotarizationInfoResult.Builder status(Status status);
-		abstract NotarizationInfoResult.Builder message(String message);
-		abstract NotarizationInfoResult.Builder notarizationLog(String notarizationLog);
-		abstract NotarizationInfoResult build();
+	public static abstract class Builder {
+		public abstract NotarizationInfoResult.Builder status(Status status);
+		public abstract NotarizationInfoResult.Builder message(String message);
+		public abstract NotarizationInfoResult.Builder notarizationLog(String notarizationLog);
+		public abstract NotarizationInfoResult build();
 	}
 }
