+++ src/main/java/org/eclipse/cbi/ws/macos/notarization/xcrun/common/SimpleStaplerResult.java
@@ -1,4 +1,4 @@
-package org.eclipse.cbi.ws.macos.notarization.xcrun.altool;
+package org.eclipse.cbi.ws.macos.notarization.xcrun.common;
 
 import com.google.auto.value.AutoValue;
